pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sfl"
version = "0.1.0"
description = "Simple Federated Learning Demo with NVFlare + Flower"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "SFL Team" }]
keywords = ["federated-learning", "nvflare", "flower", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "nvflare==2.7.1",
    "flwr[simulation]==1.17.0",
    "click>=8.1.0,<8.2.0",
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
sfl-run = "jobs.runner:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sfl"]

# ============================================================================
# Flower Configuration
# ============================================================================
[tool.flwr.app]
publisher = "sfl-demo"

[tool.flwr.app.components]
serverapp = "sfl.server.app:app"
clientapp = "sfl.client:app"

# Flower run config - these control the simulation
# Change num-server-rounds here to adjust number of rounds
# For NVFlare runner: num-clients is controlled via CLI --num-clients
[tool.flwr.app.config]
num-server-rounds = 2
num-clients = 2

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 2
address = "127.0.0.1:9093"
insecure = true

# ============================================================================
# Development Tools Configuration
# ============================================================================
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hatch
    | \.mypy_cache
    | \.venv
    | _build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", ".git"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
exclude = [".venv", "dist", "build"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
